make frontend aware of new authentication/registration/login scheme
Note: the frontend now uses a dummy username to register accounts with.
A real username input field should be added so that accounts can be
created with the correct username. Right now, if you try to register
twice, you'll get a uniqueness violation from Postgres because the dummy
username will be we used twice, or the same Facebook account will be
used twice.