add optional auth middleware; improve post listings
Using the optional auth middleware, we can make the request object of
subsequent middlewares optionally have the `account` field, just like
with the requires auth middleware. The difference here is that if no
Authorization HTTP header is present, the optional auth middleware won't
abort the request with a 401 status, it'll instead just jump to the next
middleware in the stack.

The getPosts function in the ormHelpers module now has an activeAccount
field, which it can use to personalize the post list that it produces.
In particular, if an account is active, it will check whether that
account is following each post and add that information to the post
object.

The upshot is that the frontend can now display "follow" or "unfollow"
correctly in post lists.

Closes #62