implement dummy email sending
Concerns #65

An email address can be specified in `config/gottafindit.json` as a
dummy recipient that will receive all match notifications. When matches
are made and are above the threshold set in `config/gottafindit.json`,
emails will be send via the SMTP server specified in `config/email.json`
to the dummy recipient.